requirements.txt
pillow==10.0.0 ; python_version >= "3.11" and python_version < "4.0"
